requirements.txt
# Core ML stack
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.45.0
peft>=0.12.0
bitsandbytes>=0.43.0
trl>=0.11.0
accelerate>=0.34.0
datasets>=2.21.0

# Metrics and statistics
scikit-learn>=1.5.0
statsmodels>=0.14.0
scipy>=1.13.0

# API gateway (OpenRouter is OpenAI-compatible)
openai>=1.40.0

# Data / utilities
pandas>=2.2.0
numpy>=1.26.0
Pillow>=10.4.0
tqdm>=4.66.0
pyyaml>=6.0

# Plotting
matplotlib>=3.9.0

# Manuscript rebuild
python-docx>=1.1.0
